Transfer run_iteration to Ensemble
