+++ Tests/test_metropolis.py
@@ -5,14 +5,13 @@ sys.path.append("..")
 import numpy as np
 from scipy.integrate import trapz
 from metropolis import all_signal_handler
-from metropolis import E_field, solve, select_next_params
-from metropolis import do_simulation, roll_acceptance, unpack_simpar
+from metropolis import select_next_params, check_approved_param
+from forward_solver import E_field, solve, calculate_PL, calculate_TRTS
+from metropolis import roll_acceptance
+from utils import unpack_simpar, set_min_y, search_c_grps
 from metropolis import almost_equal
-from metropolis import check_approved_param
-from metropolis import run_iteration
-from metropolis import search_c_grps
-from metropolis import set_min_y
-from sim_utils import Grid, calculate_PL, calculate_TRTS
+
+from sim_utils import Grid
 q = 1.0  # [e]
 q_C = 1.602e-19  # [C]
 kB = 8.61773e-5  # [eV / K]
@@ -410,7 +409,6 @@ class TestUtils(unittest.TestCase):
 
         np.testing.assert_almost_equal(PL_by_initvals / np.amax(PL_by_initvals), PL_by_initparams / np.amax(PL_by_initvals))
 
-
     def test_approve_param(self):
         info = {'names': ['tauP', 'tauN', 'somethingelse'],
                 'prior_dist': {'tauP': (0.1, np.inf), 'tauN': (0.1, np.inf),
@@ -689,53 +687,6 @@ class TestUtils(unittest.TestCase):
 
         return
 
-    def test_do_simulation(self):
-        # Make sure the simulation starts at 0, even if the experimental data doesn't
-        unit_conversions = {"n0": ((1e-7) ** 3), "p0": ((1e-7) ** 3),
-                            "mu_n": ((1e7) ** 2) / (1e9), "mu_p": ((1e7) ** 2) / (1e9),
-                            "ks": ((1e7) ** 3) / (1e9), "Sf": 1e-2, "Sb": 1e-2}
-
-        param_info = {"names": ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "tauN", "tauP",
-                                "Sf", "Sb", "eps", "Tm"],
-                      "active": {"n0": 0, "p0": 1,
-                                 "mu_n": 0, "mu_p": 0,
-                                 "ks": 1, "Sf": 1, "Sb": 1,
-                                 "Cn": 0, "Cp": 0,
-                                 "tauN": 1, "tauP": 1, "eps": 0, "Tm": 0},
-                      "unit_conversions": unit_conversions}
-        vals = {'n0': 1e8,
-                'p0': 3e15,
-                'mu_n': 20,
-                'mu_p': 100,
-                "ks": 1e-11,
-                "Cn": 0, "Cp": 0,
-                'tauN': 120,
-                'tauP': 200,
-                'Sf': 5,
-                'Sb': 20,
-                'eps': 1,
-                "Tm": 300}
-
-        param_info["init_guess"] = vals
-        indexes = {name: param_info["names"].index(name) for name in param_info["names"]}
-        state = [param_info["init_guess"][name] for name in param_info["names"]]
-        units = np.array([unit_conversions.get(name, 1) for name in param_info["names"]], dtype=float)
-
-        thickness = 1000
-        nx = 100
-        times = np.linspace(10, 100, 901)
-
-        iniPar = np.logspace(19, 14, nx)
-        tSteps, sol = do_simulation(state, indexes, thickness, nx, iniPar, times, hmax=4, units=units)
-        np.testing.assert_equal(tSteps[0], 0)
-
-        times = np.linspace(0, 100, 1001)
-        _, sol2 = do_simulation(state, indexes, thickness, nx, iniPar, times, hmax=4, units=units)
-        np.testing.assert_equal(sol[0], sol2[0])
-        np.testing.assert_equal(sol[-1], sol2[-1])
-
-        return
-
     def test_almost_equal(self):
         threshold = 1e-7
 
@@ -800,127 +751,6 @@ class TestUtils(unittest.TestCase):
         self.assertEqual(meas_type[2], mtype)
         return
 
-    def test_run_iter(self):
-        # Will basically need to set up a full simulation for this
-        np.random.seed(42)
-        Length = [2000, 2000]                            # Length (nm)
-        L = [2 ** 7, 2 ** 7]                                # Spatial point
-        mtype = ["TRPL", "TRPL"]
-        simPar = {"lengths": Length,
-                  "nx": L,
-                  "meas_types": mtype,
-                  "num_meas": 2}
-
-        iniPar = [1e15 * np.ones(L[0]), 1e16 * np.ones(L[1])]
-
-        param_names = ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "Tm",
-                       "Sf", "Sb", "tauN", "tauP", "eps", "m"]
-        unit_conversions = {"n0": ((1e-7) ** 3), "p0": ((1e-7) ** 3),
-                            "mu_n": ((1e7) ** 2) / (1e9), "mu_p": ((1e7) ** 2) / (1e9),
-                            "ks": ((1e7) ** 3) / (1e9), "Sf": 1e-2, "Sb": 1e-2}
-
-        # Iterations should proceed independent of which params are actively iterated,
-        # as all params are presumably needed to complete the simulation
-        param_info = {"names": param_names,
-                      "unit_conversions": unit_conversions,
-                      "active": {name: 0 for name in param_names}}
-        initial_guess = {"n0": 0,
-                         "p0": 0,
-                         "mu_n": 0,
-                         "mu_p": 0,
-                         "ks": 1e-11,
-                         "Sf": 0,
-                         "Sb": 0,
-                         "Cn": 0,
-                         "Cp": 0,
-                         "Tm": 300,
-                         "tauN": 1e99,
-                         "tauP": 1e99,
-                         "eps": 10,
-                         "m": 1}
-        param_info["init_guess"] = initial_guess
-
-        sim_flags = {"current_sigma": {"TRPL": 1},
-                     "hmax": 4, "rtol": 1e-5, "atol": 1e-8,
-                     "solver": ("solveivp",),
-                     "model": "std"}
-
-        indexes = {name: param_names.index(name) for name in param_names}
-        state = [param_info["init_guess"][name] for name in param_names]
-        units = np.array([unit_conversions.get(name, 1) for name in param_names], dtype=float)
-
-        nt = 1000
-        times = [np.linspace(0, 100, nt+1), np.linspace(0, 100, nt+1)]
-        vals = [np.ones(nt+1) * 23, np.ones(nt+1) * 23]
-        uncs = [np.ones(nt+1) * 1e-99, np.ones(nt+1) * 1e-99]
-        logll = run_iteration(state, indexes, units, simPar, iniPar, times, vals, uncs, None,
-                              sim_flags, logger=self.logger)
-
-        np.testing.assert_almost_equal(
-            logll, np.sum([-59340.105083, -32560.139058]), decimal=0)  # rtol=1e-5
-
-    def test_run_iter_depletion(self):
-        """Prove that the truncation allows likelihood of two carrier-depleting simulations to be reliably determined."""
-        np.random.seed(42)
-        Length = [2000]                            # Length (nm)
-        L = [2 ** 7]                                # Spatial point
-        mtype = ["TRPL"]
-        simPar = {"lengths": Length, "nx": L, "meas_types": mtype, "num_meas": 1}
-
-        iniPar = [1e15 * np.ones(L[0])] # PL(t=0) ~ 2e15
-
-        param_names = ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "Tm",
-                       "Sf", "Sb", "tauN", "tauP", "eps", "m"]
-        unit_conversions = {"n0": ((1e-7) ** 3), "p0": ((1e-7) ** 3),
-                            "mu_n": ((1e7) ** 2) / (1e9), "mu_p": ((1e7) ** 2) / (1e9),
-                            "ks": ((1e7) ** 3) / (1e9), "Sf": 1e-2, "Sb": 1e-2}
-
-        param_info = {"names": param_names,
-                      "unit_conversions": unit_conversions,
-                      "active": {name: 0 for name in param_names}}
-        initial_guess = {"n0": 1e8,
-                         "p0": 1e17,
-                         "mu_n": 0,
-                         "mu_p": 0,
-                         "ks": 1e-13,
-                         "Sf": 0,
-                         "Sb": 0,
-                         "Cn": 0,
-                         "Cp": 0,
-                         "Tm": 300,
-                         "tauN": 4, # Fast enough to deplete carriers
-                         "tauP": 4,
-                         "eps": 10,
-                         "m": 1}
-        param_info["init_guess"] = initial_guess
-
-        sim_flags = {"current_sigma": {"TRPL": 1},
-                     "hmax": 4, "rtol": 1e-5, "atol": 1e-8,
-                     "solver": ("solveivp",),
-                     "force_min_y": True,
-                     "model": "std"}
-
-        indexes = {name: param_names.index(name) for name in param_names}
-        state = [param_info["init_guess"][name] for name in param_names]
-        units = np.array([unit_conversions.get(name, 1) for name in param_names], dtype=float)
-
-        nt = 1000
-        times = [np.linspace(0, 100, nt+1)]
-        vals = [np.log10(2e14 * np.exp(-times[0] / 8))]
-        uncs = [np.ones(nt+1) * 1e-99]
-        logll1 = run_iteration(state, indexes, units, simPar, iniPar, times, vals, uncs, None,
-                               sim_flags, logger=self.logger)
-
-        # A small move toward the true lifetime of 10 makes the likelihood better
-        # Without min_y truncation, the likelihoods were so small they weren't even comparable
-        param_info["init_guess"]["tauN"] = 4.01
-        param_info["init_guess"]["tauP"] = 4.01
-
-        state = [param_info["init_guess"][name] for name in param_names]
-        logll2 = run_iteration(state, indexes, units, simPar, iniPar, times, vals, uncs, None,
-                               sim_flags, logger=self.logger)
-        self.assertTrue(logll2 > logll1)
-
     def test_set_min_y(self):
         t = np.linspace(0, 100, 100)
         vals = np.log10(np.exp(-t / 2)) # A slow decay
@@ -932,262 +762,10 @@ class TestUtils(unittest.TestCase):
         np.testing.assert_equal(sol[len(sol)-n_set:], min_y)
         self.assertEqual(np.log10(min_y), min(vals) - scale_shift)
 
-    def test_run_iter_cutoff(self):
-        # Same as test_run_iter, only "experimental" data is
-        # truncated at [50,100] instead of [0,100].
-        # Half as many points means the likelihood should be reduced to about half.
-        np.random.seed(42)
-        Length = [2000, 2000]                            # Length (nm)
-        L = [2 ** 7, 2 ** 7]                                # Spatial point
-        mtype = ["TRPL", "TRPL"]
-        simPar = {"lengths": Length,
-                  "nx": L,
-                  "meas_types": mtype,
-                  "num_meas": 2}
-
-        iniPar = [1e15 * np.ones(L[0]), 1e16 * np.ones(L[1])]
-
-        param_names = ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "Tm",
-                       "Sf", "Sb", "tauN", "tauP", "eps", "m"]
-        unit_conversions = {"n0": ((1e-7) ** 3), "p0": ((1e-7) ** 3),
-                            "mu_n": ((1e7) ** 2) / (1e9), "mu_p": ((1e7) ** 2) / (1e9),
-                            "ks": ((1e7) ** 3) / (1e9), "Sf": 1e-2, "Sb": 1e-2}
-
-        # Iterations should proceed independent of which params are actively iterated,
-        # as all params are presumably needed to complete the simulation
-        param_info = {"names": param_names,
-                      "unit_conversions": unit_conversions,
-                      "active": {name: 0 for name in param_names}}
-        initial_guess = {"n0": 0,
-                         "p0": 0,
-                         "mu_n": 0,
-                         "mu_p": 0,
-                         "ks": 1e-11,
-                         "Sf": 0,
-                         "Sb": 0,
-                         "Cn": 0,
-                         "Cp": 0,
-                         "Tm": 300,
-                         "tauN": 1e99,
-                         "tauP": 1e99,
-                         "eps": 10,
-                         "m": 1}
-        param_info["init_guess"] = initial_guess
-
-        sim_flags = {"current_sigma": {"TRPL": 1},
-                     "hmax": 4, "rtol": 1e-5, "atol": 1e-8,
-                     "solver": ("solveivp",),
-                     "model": "std"}
-
-        indexes = {name: param_names.index(name) for name in param_names}
-        state = [param_info["init_guess"][name] for name in param_names]
-        units = np.array([unit_conversions.get(name, 1) for name in param_names], dtype=float)
-
-        nt = 500
-        times = [np.linspace(50, 100, nt+1), np.linspace(50, 100, nt+1)]
-        vals = [np.ones(nt+1) * 23, np.ones(nt+1) * 23]
-        uncs = [np.ones(nt+1) * 1e-99, np.ones(nt+1) * 1e-99]
-        logll = run_iteration(state, indexes, units, simPar, iniPar, times, vals, uncs, None,
-                              sim_flags, logger=self.logger)
-
-        # First iter; auto-accept
-        np.testing.assert_almost_equal(
-            logll, np.sum([-29701, -16309]), decimal=0)  # rtol=1e-5
-
-
-    def test_run_iter_scale(self):
-        # Same as test_run_iter, except global scale factor, which will be chosen to match
-        # the first measurement (but off by one order of magnitude from the second)
-        # likelihood should be small but nonzero as a result
-        np.random.seed(42)
-        Length = [2000, 2000]                            # Length (nm)
-        L = [2 ** 7, 2 ** 7]                                # Spatial point
-        mtype = ["TRPL", "TRPL"]
-        simPar = {"lengths": Length,
-                  "nx": L,
-                  "meas_types": mtype,
-                  "num_meas": 2}
-
-        iniPar = [1e15 * np.ones(L[0]), 1e16 * np.ones(L[1])]
-
-        param_names = ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "Tm",
-                       "Sf", "Sb", "tauN", "tauP", "eps"]
-        unit_conversions = {"n0": ((1e-7) ** 3), "p0": ((1e-7) ** 3),
-                            "mu_n": ((1e7) ** 2) / (1e9), "mu_p": ((1e7) ** 2) / (1e9),
-                            "ks": ((1e7) ** 3) / (1e9), "Sf": 1e-2, "Sb": 1e-2}
-
-        # Iterations should proceed independent of which params are actively iterated,
-        # as all params are presumably needed to complete the simulation
-        param_info = {"names": param_names,
-                      "unit_conversions": unit_conversions,
-                      "active": {name: 0 for name in param_names}}
-        initial_guess = {"n0": 0,
-                         "p0": 0,
-                         "mu_n": 0,
-                         "mu_p": 0,
-                         "ks": 1e-20,
-                         "Sf": 0,
-                         "Sb": 0,
-                         "Cn": 0,
-                         "Cp": 0,
-                         "Tm": 300,
-                         "tauN": 1e99,
-                         "tauP": 1e99,
-                         "eps": 10}
-        param_info["init_guess"] = initial_guess
-
-        sim_flags = {"current_sigma": {"TRPL": 1},
-                     "hmax": 4, "rtol": 1e-5, "atol": 1e-8,
-                     "scale_factor": (0.02, [0, 1, 2, 3, 4, 5], [(0, 2, 4), (1, 3, 5)]),
-                     "solver": ("solveivp",),
-                     "model": "std"}
-
-        # These would normally be inserted when the script file is read by bayes_io
-        param_info["names"].append("_s0")
-        param_info["names"].append("_s1")
-        # By setting individual scale factors in this simple case the likelihood can be made perfect
-        param_info["init_guess"]["_s0"] = 2e-17 ** -1
-        param_info["init_guess"]["_s1"] = 2e-15 ** -1
-        indexes = {name: param_names.index(name) for name in param_names}
-        state = [param_info["init_guess"][name] for name in param_names]
-        units = np.array([unit_conversions.get(name, 1) for name in param_names], dtype=float)
-
-        nt = 1000
-        times = [np.linspace(0, 100, nt+1), np.linspace(0, 100, nt+1)]
-        vals = [np.ones(nt+1) * 23, np.ones(nt+1) * 23]
-        uncs = [np.ones(nt+1) * 1e-99, np.ones(nt+1) * 1e-99]
-
-        logll = run_iteration(state, indexes, units, simPar, iniPar, times, vals, uncs, None,
-                              sim_flags, logger=self.logger)
-
-        np.testing.assert_almost_equal(
-            logll, 0, decimal=0)  # rtol=1e-5
-
-    def test_run_iter_mixed_types(self):
-        # Will basically need to set up a full simulation for this
-        np.random.seed(42)
-        Length = [2000, 2000]                            # Length (nm)
-        L = [2 ** 7, 2 ** 7]                                # Spatial point
-        mtype = ["TRPL", "TRTS"]
-        simPar = {"lengths": Length,
-                  "nx": L,
-                  "meas_types": mtype,
-                  "num_meas": 2}
-
-        iniPar = [1e15 * np.ones(L[0]), 1e15 * np.ones(L[1])]
-
-        param_names = ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "Tm",
-                       "Sf", "Sb", "tauN", "tauP", "eps", "m"]
-        unit_conversions = {"n0": ((1e-7) ** 3), "p0": ((1e-7) ** 3),
-                            "mu_n": ((1e7) ** 2) / (1e9), "mu_p": ((1e7) ** 2) / (1e9),
-                            "ks": ((1e7) ** 3) / (1e9), "Sf": 1e-2, "Sb": 1e-2}
-
-        # Iterations should proceed independent of which params are actively iterated,
-        # as all params are presumably needed to complete the simulation
-        param_info = {"names": param_names,
-                      "unit_conversions": unit_conversions,
-                      "active": {name: 0 for name in param_names}}
-        initial_guess = {"n0": 0,
-                         "p0": 0,
-                         "mu_n": 0.01,
-                         "mu_p": 0.01,
-                         "ks": 1e-11,
-                         "Sf": 0,
-                         "Sb": 0,
-                         "Cn": 0,
-                         "Cp": 0,
-                         "Tm": 300,
-                         "tauN": 1e99,
-                         "tauP": 1e99,
-                         "eps": 10,
-                         "m": 1}
-        param_info["init_guess"] = initial_guess
-
-        sim_flags = {"current_sigma": {"TRPL": 1, "TRTS": 10},
-                     "hmax": 4, "rtol": 1e-5, "atol": 1e-8,
-                     "solver": ("solveivp",),
-                     "model": "std"}
-
-        indexes = {name: param_names.index(name) for name in param_names}
-        state = [param_info["init_guess"][name] for name in param_names]
-        units = np.array([unit_conversions.get(name, 1) for name in param_names], dtype=float)
-
-        nt = 1000
-        times = [np.linspace(0, 100, nt+1), np.linspace(0, 100, nt+1)]
-        vals = [np.ones(nt+1) * 23, np.ones(nt+1) * -2]
-        uncs = [np.ones(nt+1) * 1e-99, np.ones(nt+1) * 1e-99]
-        logll = run_iteration(state, indexes, units, simPar, iniPar, times, vals, uncs, None,
-                              sim_flags, logger=self.logger)
-
-        # First iter; auto-accept
-        np.testing.assert_almost_equal(
-            logll, np.sum([-59340.105083, -517.98]), decimal=0)  # rtol=1e-5
-
     def test_one_sim_ll_errata(self):
         # TODO: The next time odeint fails to do a simulation, upload it into this
         # test case
         return
-        # np.random.seed(42)
-        # Length  = 2000                            # Length (nm)
-        # L   = 2 ** 7                                # Spatial points
-        # plT = 1                                  # Set PL interval (dt)
-        # pT  = (0,1,3,10,30,100)                   # Set plot intervals (%)
-        # tol = 7                                   # Convergence tolerance
-        # MAX = 10000                                  # Max iterations
-
-        # simPar = [Length, -1, L, -1, plT, pT, tol, MAX]
-
-        # iniPar = [np.logspace(20,1,L), 1e16 * np.ones(L)]
-
-        # param_names = ["n0", "p0", "mu_n", "mu_p", "ks", "Cn", "Cp", "Tm",
-        #                "Sf", "Sb", "tauN", "tauP", "eps", "m"]
-        # unit_conversions = {"n0":((1e-7) ** 3), "p0":((1e-7) ** 3),
-        #                     "mu_n":((1e7) ** 2) / (1e9), "mu_p":((1e7) ** 2) / (1e9),
-        #                     "ks":((1e7) ** 3) / (1e9), "Sf":1e-2, "Sb":1e-2}
-
-        # # Iterations should proceed independent of which params are actively iterated,
-        # # as all params are presumably needed to complete the simulation
-        # param_info = {"names":param_names,
-        #               "unit_conversions":unit_conversions,
-        #               "active":{name:0 for name in param_names}}
-        # initial_guess = {"n0":1e8,
-        #                  "p0":1e12,
-        #                  "mu_n":2,
-        #                  "mu_p":2,
-        #                  "ks":1e-11,
-        #                  "Sf":1000,
-        #                  "Sb":1e4,
-        #                  "Cn":0,
-        #                  "Cp":0,
-        #                  "Tm":300,
-        #                  "tauN":10,
-        #                  "tauP":10,
-        #                  "eps":10,
-        #                  "m":0}
-
-        # sim_flags = {"hmax":MIN_HMAX * 4, "rtol":1e-10, "atol":1e-10,
-        #              "measurement":"TRPL",
-        #              "solver":("odeint",)}
-
-        # nt = 100
-        # i = 0
-        # times = [np.linspace(0, 10, nt+1), np.linspace(0, 10, nt+1)]
-        # vals = [np.zeros(nt+1), np.zeros(nt+1)]
-
-        # running_hmax = [sim_flags["hmax"]] * len(iniPar)
-        # p = Parameters(param_info, initial_guess)
-        # p.apply_unit_conversions(param_info)
-
-        # with self.assertLogs() as captured:
-        #     one_sim_likelihood(p, simPar, running_hmax, sim_flags, self.logger,
-        #                        (i, iniPar[i], times[i], vals[i]))
-
-        # # We can force a simulation failure by using odeint + rapid recombination
-        # self.assertEqual(len(captured.records), 8)
-        # self.assertEqual(captured.records[1].getMessage(),
-        #                  f"{i}: Carriers depleted!")
-        # self.assertEqual(captured.records[2].getMessage(),
-        #                  f"{i}: Retrying hmax={MIN_HMAX * 2}")
 
 if __name__ == "__main__":
     t = TestUtils()
