+++ metropolis.py
@@ -5,232 +5,19 @@ Created on Mon Jan 31 22:13:26 2022
 @author: cfai2
 """
 import os
-import sys
 import signal
 import pickle
 import numpy as np
-from scipy.integrate import solve_ivp, odeint
 
-from forward_solver import MODELS
-from sim_utils import Ensemble, Grid, Solution
+from sim_utils import Ensemble
 from mcmc_logging import start_logging, stop_logging
 from bayes_io import make_dir
-from laplace import make_I_tables, do_irf_convolution, post_conv_trim
-try:
-    from nn_features import NeuralNetwork
-    HAS_NN_LIB = True
-    nn = NeuralNetwork()
-except ImportError:
-    HAS_NN_LIB = False
+from laplace import make_I_tables
 
 # Constants
-eps0 = 8.854 * 1e-12 * 1e-9  # [C / V m] to {C / V nm}
-q = 1.0  # [e]
-q_C = 1.602e-19  # [C]
-kB = 8.61773e-5  # [eV / K]
-DEFAULT_RTOL = 1e-7
-DEFAULT_ATOL = 1e-10
-DEFAULT_HMAX = 4
 MAX_PROPOSALS = 100
 MSG_FREQ = 100
 MSG_COOLDOWN = 3 # Log first few states regardless of verbose
-
-# Allow this proportion of simulation values to become negative due to convolution,
-# else the simulation is failed.
-NEGATIVE_FRAC_TOL = 0.2
-
-def U(x):
-    return 1000 * (x < -2) + 1 * (1 + np.sin(2*np.pi*x)) * np.logical_and(-2 <= x, x <= -1.25) \
-                             + 2 * (1 + np.sin(2*np.pi*x)) * np.logical_and(-1.25 <= x, x <= -0.25) \
-                             + 3 * (1 + np.sin(2*np.pi*x)) * np.logical_and(-0.25 <= x, x <= 0.75) \
-                             + 4 * (1 + np.sin(2*np.pi*x)) * np.logical_and(0.75 <= x, x <= 1.75) \
-                             + 5 * (1 + np.sin(2*np.pi*x)) * np.logical_and(1.75 <= x, x <= 2) \
-                             + 1000 * (x > 2)
-
-
-def E_field(N, P, n0, p0, eps, dx, corner_E=0):
-    if N.ndim == 1:
-        E = corner_E + q_C / (eps * eps0) * dx * \
-            np.cumsum(((P - p0) - (N - n0)))
-        E = np.concatenate(([corner_E], E))
-    elif N.ndim == 2:
-        E = corner_E + q_C / (eps * eps0) * dx * \
-            np.cumsum(((P - p0) - (N - n0)), axis=1)
-        num_tsteps = len(N)
-        E = np.concatenate((np.ones(shape=(num_tsteps, 1))*corner_E, E), axis=1)
-    return E
-
-
-def solve(iniPar, g, state, indexes, meas="TRPL", units=None, solver=("solveivp",), model="std",
-          RTOL=DEFAULT_RTOL, ATOL=DEFAULT_ATOL):
-    """
-    Calculate one simulation. Outputs in same units as measurement data,
-    ([cm, V, s]) for PL.
-
-    Parameters
-    ----------
-    iniPar : np.ndarray
-        Initial conditions - either an array of one initial value per g.nx, or an array
-        of parameters (e.g. [fluence, alpha, direction]) usable to generate the initial condition.
-    g : Grid
-        Object containing space and time grid information.
-    state : ndarray
-        An array of parameters, ordered according to param_info["names"],
-        corresponding to a state in the parameter space.
-    indexes : dict[str, int]
-        A map of parameter names and their corresponding indices in the state array.
-    meas : str, optional
-        Type of measurement (e.g. TRPL, TRTS) being simulated. The default is "TRPL".
-    units : dict[str], optional
-        Unit conversions to be applied to each parameter. The default is None.
-    solver : tuple(str), optional
-        Solution method used to perform simulation and optional related args.
-        The first element is the solver type.
-        Choices include:
-        solveivp - scipy.integrate.solve_ivp()
-        odeint - scipy.integrate.odeint()
-        NN - a tensorflow/keras model (WIP!)
-        All subsequent elements are optional. For NN the second element is
-        a path to the NN weight file, the third element is a path
-        to the corresponding NN scale factor file.
-        The default is ("solveivp",).
-    model : str, optional
-        Physics model to be solved by the solver, chosen from MODELS.
-        The default is "std".
-    RTOL, ATOL : float, optional
-        Tolerance parameters for scipy solvers. See the solve_ivp() docs for details.
-
-    Returns
-    -------
-    sol : np.ndarray
-        Array of values (e.g. TRPL) from final simulation.
-    next_init : np.ndarray
-        Values (e.g. the electron profile) at the final time of the simulation.
-
-    """
-    if units is None:
-        units = np.ones_like(state)
-    if solver[0] == "solveivp" or solver[0] == "odeint" or solver[0] == "diagnostic":
-        if len(iniPar) == g.nx:         # If list of initial values
-            init_dN = iniPar * 1e-21    # [cm^-3] to [nm^-3]
-        else:                           # List of parameters
-            fluence = iniPar[0] * 1e-14 # [cm^-2] to [nm^-2]
-            alpha = iniPar[1] * 1e-7    # [cm^-1] to [nm^-1]
-            init_dN = fluence * alpha * np.exp(-alpha * g.xSteps)
-            try:
-                init_dN = init_dN[::np.sign(int(iniPar[2]))]
-            except (IndexError, ValueError):
-                pass
-
-        state *= units
-        N = init_dN + state[indexes["n0"]]
-        P = init_dN + state[indexes["p0"]]
-        E_f = E_field(N, P, state[indexes["n0"]], state[indexes["p0"]], state[indexes["eps"]], g.dx)
-
-
-        # Depends on how many dependent variables and parameters are in the selected model
-        if model == "std":
-            init_condition = np.concatenate([N, P, E_f], axis=None)
-            args = (g.nx, g.dx, state[indexes["n0"]], state[indexes["p0"]], state[indexes["mu_n"]], state[indexes["mu_p"]],
-                    state[indexes["ks"]], state[indexes["Cn"]], state[indexes["Cp"]],
-                    state[indexes["Sf"]], state[indexes["Sb"]], state[indexes["tauN"]], state[indexes["tauP"]],
-                    ((q_C) / (state[indexes["eps"]] * eps0)), state[indexes["Tm"]],)
-        elif model == "traps":
-            init_condition = np.concatenate([N, np.zeros_like(N), P, E_f], axis=None)
-            args = (g.nx, g.dx, state[indexes["n0"]], state[indexes["p0"]], state[indexes["mu_n"]], state[indexes["mu_p"]],
-                    state[indexes["ks"]], state[indexes["Cn"]], state[indexes["Cp"]],
-                    state[indexes["Sf"]], state[indexes["Sb"]], state[indexes["tauN"]], state[indexes["tauP"]],
-                    ((q_C) / (state[indexes["eps"]] * eps0)), state[indexes["Tm"]],
-                    state[indexes["kC"]], state[indexes["Nt"]], state[indexes["tauE"]],)
-        else:
-            raise ValueError(f"Invalid model {model}")
-
-        dy = lambda t, y: MODELS[model](t, y, *args)
-
-        s = Solution()
-        # Can't get solve_ivp's root finder to work reliably, so leaving out the early termination events for now.
-        # Instead we will let solve_ivp run as-is, and then find where the early termination cutoff would have been
-        # after the fact.
-        # if meas == "TRPL":
-        #     min_y = g.min_y * 1e-23 # To nm/ns units
-        #     stop_integrate = lambda t, y: check_threshold(t, y, g.nx, g.dx, min_y=min_y, mode="TRPL",
-        #                                                   ks=p.ks, n0=p.n0, p0=p.p0)
-        #     stop_integrate.terminal = 1
-
-        # elif meas == "TRTS":
-        #     min_y = g.min_y * 1e-9
-        #     stop_integrate = lambda t, y: check_threshold(t, y, g.nx, g.dx, min_y=min_y, mode="TRTS",
-        #                                                   mu_n=p.mu_n, mu_p=p.mu_p, n0=p.n0, p0=p.p0)
-        #     stop_integrate.terminal = 1
-        # else:
-        #     raise NotImplementedError("TRPL or TRTS only")
-
-        i_final = len(g.tSteps)
-        if solver[0] == "solveivp" or solver[0] == "diagnostic":
-            sol = solve_ivp(dy, [g.start_time, g.time], init_condition,
-                            method='LSODA', dense_output=True, # events=stop_integrate,
-                            max_step=g.hmax, rtol=RTOL, atol=ATOL)
-
-            data = sol.sol(g.tSteps).T
-            data[g.tSteps > sol.t[-1]] = 0 # Disallow sol from extrapolating beyond time it solved up to
-            # if len(sol.t_events[0]) > 0:
-            #     t_final = sol.t_events[0][0]
-            #     try:
-            #         i_final = np.where(g.tSteps < t_final)[0][-1]
-            #     except IndexError:
-            #         pass
-
-        else:
-            data = odeint(MODELS[model], init_condition, g.tSteps, args=args,
-                      hmax=g.hmax, rtol=RTOL, atol=ATOL, tfirst=True)
-
-        # Also depends on how many dependent variables
-        if model == "std":
-            s.N, s.P, E_f = np.split(data, [g.nx, 2*g.nx], axis=1)
-        elif model == "traps":
-            s.N, s.N_trap, s.P, E_f = np.split(data, [g.nx, 2*g.nx, 3*g.nx], axis=1)
-
-        if meas == "TRPL":
-            s.calculate_PL(g, state[indexes["ks"]], state[indexes["n0"]], state[indexes["p0"]])
-            state /= units  # [nm, V, ns] to [cm, V, s]
-            s.PL *= 1e23                            # [nm^-2 ns^-1] to [cm^-2 s^-1]
-            i_final = np.argmax(s.PL < g.min_y)
-            if s.PL[i_final] < g.min_y:
-                s.PL[i_final:] = g.min_y
-            return s.PL
-        elif meas == "TRTS":
-            s.calculate_TRTS(g, state[indexes["mu_n"]], state[indexes["mu_p"]], state[indexes["n0"]], state[indexes["p0"]])
-            state /= units
-            s.trts *= 1e9
-            i_final = np.argmax(s.trts < g.min_y)
-            if s.trts[i_final] < g.min_y:
-                s.trts[i_final:] = g.min_y
-            return s.trts
-        else:
-            raise NotImplementedError("TRTS or TRPL only")
-
-    elif solver[0] == "NN":
-        if not HAS_NN_LIB:
-            raise ImportError("Failed to load neural network library")
-
-        if meas != "TRPL":
-            raise NotImplementedError("TRPL only")
-
-        if not nn.has_model:
-            nn.load_model(solver[1], solver[2])
-
-        scaled_matPar = np.zeros((1, 14))
-        scaled_matPar[0] = [state[indexes["p0"]], state[indexes["mu_n"]], state[indexes["mu_p"]],
-                            state[indexes["ks"]], state[indexes["Cn"]], state[indexes["Cp"]],
-                            state[indexes["Sf"]], state[indexes["Sb"]], state[indexes["tauN"]], state[indexes["tauP"]],
-                            state[indexes["eps"]]**-1,
-                            iniPar[0], iniPar[1], g.thickness]
-
-        pl_from_NN = nn.predict(g.tSteps, scaled_matPar)
-        return pl_from_NN
-
-    else:
-        raise NotImplementedError
 
 
 def check_approved_param(new_state, param_info, indexes, active, do_log):
@@ -333,102 +120,6 @@ def select_next_params(current_state, param_info, indexes, is_active, trial_move
     return new_state
 
 
-def do_simulation(state, indexes, thickness, nx, iniPar, times, hmax, meas="TRPL", units=None,
-                  solver=("solveivp",), model="std", rtol=DEFAULT_RTOL, atol=DEFAULT_ATOL):
-    """ Set up one simulation. """
-    g = Grid()
-    g.thickness = thickness
-    g.nx = nx
-    g.dx = g.thickness / g.nx
-    g.xSteps = np.linspace(g.dx / 2, g.thickness - g.dx/2, g.nx)
-
-    g.time = times[-1]
-    g.start_time = 0
-    g.nt = len(times) - 1
-    g.hmax = hmax
-    g.tSteps = times
-
-    if times[0] > 0:
-        # Always start sims from t=0 even if experimental data doesn't, to verify initial conditions
-        dt_estimate = times[1] - times[0]
-        g.tSteps = np.concatenate((np.arange(0, times[0], dt_estimate), g.tSteps))
-
-    sol = solve(
-        iniPar, g, state, indexes, meas=meas, units=units, solver=solver, model=model, RTOL=rtol, ATOL=atol)
-    return g.tSteps, sol
-
-
-def converge_simulation(i, state, indexes, sim_info, iniPar, times, vals, units,
-                        MCMC_fields, logger=None, verbose=True):
-    """
-    Retest and repeat simulation until all stipulated convergence criteria
-    are met.
-    For some models, this can be as simple as running do_simulation()
-    or equivalent once.
-
-    Parameters
-    ----------
-    i : int
-        Index of ith simulation in a measurement set requiring n simulations.
-    state : ndarray
-        An array of parameters, ordered according to param_info["names"],
-        corresponding to a state in the parameter space.
-    indexes : dict[str, int]
-        A map of parameter names and their corresponding indices in the state array.
-    sim_info : dict
-        Dictionary compatible with unpack_simpar(),
-        containing a thickness, nx, and measurement type info
-        needed for simulation.
-    iniPar : ndarray
-        Array of initial conditions for simulation.
-    times : ndarray
-        Time points at which to evaluate the simulation.
-    vals : ndarray
-        Actual values to compare simulation output against.
-    MCMC_fields : dict
-        Dictionary of MMC control parameters.
-    logger : logger
-        Logger to write status messages to. The default is None.
-    verbose : bool, optional
-        Print more detailed status messages. The default is False.
-
-    Returns
-    -------
-    sol : ndarray
-        Array of values (e.g. TRPL) from final simulation.
-    tSteps : ndarray
-        Array of times the final simulation was evaluated at.
-    success : bool
-        Whether the final simulation passed all convergence criteria.
-
-    """
-    success = True
-    thickness, nx, meas_type = unpack_simpar(sim_info, i)
-
-    rtol = MCMC_fields.get("rtol", DEFAULT_RTOL)
-    atol = MCMC_fields.get("atol", DEFAULT_ATOL)
-    hmax = MCMC_fields.get("hmax", DEFAULT_HMAX)
-
-    t_steps = np.array(times)
-    sol = np.zeros_like(t_steps)
-
-    try:
-        t_steps, sol = do_simulation(state, indexes, thickness, nx, iniPar, times, hmax=hmax,
-                                    meas=meas_type, units=units,
-                                    solver=MCMC_fields["solver"], model=MCMC_fields["model"],
-                                    rtol=rtol, atol=atol)
-    except ValueError as e:
-        success = False
-        if logger is not None:
-            logger.warning(f"{i}: Simulation error occurred: {e}")
-        return t_steps, sol, success
-
-    if verbose and logger is not None:
-        logger.info(f"{i}: Simulation complete hmax={hmax}; t {t_steps[0]}-{t_steps[-1]}; x {thickness}")
-
-    return t_steps, sol, success
-
-
 def roll_acceptance(logratio):
     accepted = False
     if logratio >= 0: # Automatic accept
@@ -441,184 +132,11 @@ def roll_acceptance(logratio):
     return accepted
 
 
-def unpack_simpar(sim_info, i):
-    thickness = sim_info["lengths"][i]
-    nx = sim_info["nx"][i]
-    meas_type = sim_info["meas_types"][i]
-    return thickness, nx, meas_type
-
-
-def search_c_grps(c_grps : list[tuple], i : int) -> int:
-    """
-    Find the constraint group that contains i
-    and return its first value
-    """
-    for c_grp in c_grps:
-        for c in c_grp:
-            if i == c:
-                return c_grp[0]
-    return i
-
-
-def set_min_y(sol, vals, scale_shift):
-    """
-    Raise the values in (sol + scale_shift) to at least the minimum of vals.
-    scale_shift and vals should be in log scale; sol in regular scale
-    Returns:
-    sol : np.ndarray
-        New sol with raised values.
-    min_y : float
-        min_val sol was raised to. Regular scale.
-    n_set : int
-        Number of values in sol raised.
-    """
-    min_y = 10 ** min(vals - scale_shift)
-    i_final = np.searchsorted(-sol, -min_y)
-    sol[i_final:] = min_y
-    return sol, min_y, len(sol[i_final:])
-
-
 def almost_equal(x, x0, threshold=1e-10):
     if x.shape != x0.shape:
         return False
 
     return np.abs(np.nanmax((x - x0) / x0)) < threshold
-
-
-def one_sim_likelihood(state, indexes, units, sim_info, IRF_tables, MCMC_fields, logger, verbose, args):
-    i, iniPar, times, vals, uncs = args
-    meas_type = sim_info["meas_types"][i]
-    irf_convolution = MCMC_fields.get("irf_convolution", None)
-
-    ff = MCMC_fields.get("fittable_fluences", None)
-    if (ff is not None and i in ff[1]):
-        if ff[2] is not None and len(ff[2]) > 0:
-            name = f"_f{search_c_grps(ff[2], i)}"
-        else:
-            name = f"_f{i}"
-        iniPar[0] *= state[indexes[name]]
-    fa = MCMC_fields.get("fittable_absps", None)
-    if (fa is not None and i in fa[1]):
-        if fa[2] is not None and len(fa[2]) > 0:
-            name = f"_a{search_c_grps(fa[2], i)}"
-        else:
-            name = f"_a{i}"
-        iniPar[1] *= state[indexes[name]]
-    fs = MCMC_fields.get("scale_factor", None)
-    if (fs is not None and i in fs[1]):
-        if fs[2] is not None and len(fs[2]) > 0:
-            name = f"_s{search_c_grps(fs[2], i)}"
-        else:
-            name = f"_s{i}"
-        scale_shift = np.log10(state[indexes[name]])
-    else:
-        scale_shift = 0
-
-    if meas_type == "pa":
-        tSteps = np.array([0])
-        sol = np.array([U(state[0])])
-        success = True
-    else:
-        tSteps, sol, success = converge_simulation(i, state, indexes, sim_info, iniPar, times, vals, units,
-                                                   MCMC_fields, logger, verbose)
-    if not success:
-        likelihood = -np.inf
-        return likelihood
-
-    try:
-        if irf_convolution is not None and irf_convolution[i] != 0:
-            if verbose and logger is not None:
-                logger.info(f"Convolving with wavelength {irf_convolution[i]}")
-            wave = int(irf_convolution[i])
-            tSteps, sol, success = do_irf_convolution(
-                tSteps, sol, IRF_tables[wave], time_max_shift=True)
-            if not success:
-                raise ValueError("Conv failed. Check measurement data times for floating-point inaccuracies.\n"
-                                 "This may also happen if simulated signal decays extremely slowly.")
-            sol, times_c, vals_c, uncs_c = post_conv_trim(tSteps, sol, times, vals, uncs)
-
-        else:
-            # Still need to trim, in case experimental data doesn't start at t=0
-            times_c = times
-            vals_c = vals
-            uncs_c = uncs
-            sol = sol[-len(times_c):]
-
-    except ValueError as e:
-        logger.warning(e)
-        likelihood = -np.inf
-        return likelihood
-
-    if verbose and logger is not None:
-        logger.info(f"Comparing times {times_c[0]}-{times_c[-1]}")
-
-    try:
-        # TRPL must be positive!
-        # Any simulation which results in depleted carrier is clearly incorrect
-        # A few negative values may also be introduced during convolution -
-        # so we want to tolerate these, while too many suggests that depletion
-        # is happening instead
-
-        where_failed = sol < 0
-        n_fails = np.sum(where_failed)
-        success = n_fails < NEGATIVE_FRAC_TOL * len(sol)
-        if not success:
-            raise ValueError(f"{i}: Simulation failed: too many negative vals")
-
-        if n_fails > 0:
-            logger.warning(f"{i}: {n_fails} / {len(sol)} non-positive vals")
-
-        sol[where_failed] *= -1
-    except ValueError as e:
-        logger.warning(e)
-        likelihood = -np.inf
-        return likelihood
-
-    if MCMC_fields.get("force_min_y", False):
-        sol, min_y, n_set = set_min_y(sol, vals_c, scale_shift)
-        if verbose and logger is not None:
-            logger.warning(f"min_y: {min_y}")
-            if n_set > 0:
-                logger.warning(f"{n_set} values raised to min_y")
-
-    if meas_type == "pa":
-        likelihood = -sol[0]
-    else:
-        try:
-            err_sq = (np.log10(sol) + scale_shift - vals_c) ** 2
-
-            # Compatibility with single sigma
-            if isinstance(MCMC_fields["current_sigma"], dict):
-                likelihood = - \
-                    np.sum(err_sq / (MCMC_fields["current_sigma"][meas_type]**2 + 2*uncs_c**2))
-            else:
-                likelihood = - \
-                    np.sum(err_sq / (MCMC_fields["current_sigma"]**2 + 2*uncs_c**2))
-
-            if np.isnan(likelihood):
-                raise ValueError(f"{i}: Simulation failed: invalid likelihood")
-        except ValueError as e:
-            logger.warning(e)
-            likelihood = -np.inf
-    return likelihood
-
-
-def run_iteration(state, indexes, units, sim_info, iniPar, times, vals, uncs, IRF_tables,
-                  MCMC_fields, logger, verbose=False):
-    """
-    Calculates likelihood of a new proposed state
-    """
-
-    logll = np.zeros(sim_info["num_meas"])
-
-    for i in range(sim_info["num_meas"]):
-        logll[i] = one_sim_likelihood(
-            state, indexes, units, sim_info, IRF_tables, MCMC_fields, logger, verbose,
-            (i, np.array(iniPar[i]), times[i], vals[i], uncs[i]))
-        
-    logll = np.sum(logll)
-
-    return logll
 
 
 def main_metro_loop(MS_list : Ensemble, starting_iter, num_iters,
@@ -659,9 +177,7 @@ def main_metro_loop(MS_list : Ensemble, starting_iter, num_iters,
             logger.info("Simulating initial state:")
         # Calculate likelihood of initial guess
         for MS in MS_list.MS:
-            logll = run_iteration(MS.init_state, MS_list.param_indexes, MS_list.ensemble_fields["units"], MS_list.sim_info, MS_list.iniPar,
-                                  MS_list.times, MS_list.vals, MS_list.uncs, MS_list.IRF_tables,
-                                  MS.MCMC_fields, logger, verbose)
+            logll = MS_list.eval_trial_move(MS.init_state, MS.MCMC_fields, logger, verbose)
 
             MS.H.loglikelihood[0, 0] = logll
             MS.H.states[:, 0] = MS.init_state
@@ -718,9 +234,7 @@ def main_metro_loop(MS_list : Ensemble, starting_iter, num_iters,
                     if (verbose or k % MSG_FREQ == 0 or k < starting_iter + MSG_COOLDOWN) and logger is not None:
                         MS.print_status(k - 1, MS_list.ensemble_fields["active"], new_state, logger)
 
-                    logll = run_iteration(new_state, MS_list.param_indexes, MS_list.ensemble_fields["units"], MS_list.sim_info, MS_list.iniPar,
-                                          MS_list.times, MS_list.vals, MS_list.uncs, MS_list.IRF_tables,
-                                          MS.MCMC_fields, logger, verbose)
+                    logll = MS_list.eval_trial_move(new_state, MS.MCMC_fields, logger, verbose)
                     
                     if verbose and logger is not None:
                         logger.info(f"Log likelihood of proposed move: {MS.MCMC_fields.get('_beta', 1) * logll}")
